Rename blockType directive file
